[DDW-821] Consolidates Settings helper functions
